+++ packages/server/api/paths/resources.js
@@ -1,5 +1,5 @@
 export default {
-  '/api/apps/{appId}/resources/{resourceType}': {
+  '/apps/{appId}/resources/{resourceType}': {
     parameters: [
       { $ref: '#/components/parameters/appId' },
       { $ref: '#/components/parameters/resourceType' },
@@ -42,7 +42,7 @@ export default {
       security: [{ studio: [] }, {}],
     },
   },
-  '/api/apps/{appId}/resources/{resourceType}/{resourceId}': {
+  '/apps/{appId}/resources/{resourceType}/{resourceId}': {
     parameters: [
       { $ref: '#/components/parameters/appId' },
       { $ref: '#/components/parameters/resourceType' },