+++ packages/studio/components/App/App.jsx
@@ -14,6 +14,7 @@ import Login from '../Login';
 import Message from '../Message';
 import OAuth2Connect from '../OAuth2Connect';
 import OrganizationInvite from '../OrganizationInvite';
+import OrganizationProvider from '../OrganizationProvider';
 import ProtectedRoute from '../ProtectedRoute';
 import Register from '../Register';
 import ResetPassword from '../ResetPassword';
@@ -27,28 +28,30 @@ export default function App() {
     <IntlProvider defaultLocale="en-US" locale="en-US">
       <BrowserRouter>
         <UserProvider>
-          <ErrorHandler fallback={ErrorFallback}>
-            <Helmet defaultTitle="Appsemble" titleTemplate="Appsemble · %s" />
-            <Toolbar />
-            <Switch>
-              <Route component={AppList} exact path="/apps" />
-              <ProtectedRoute component={Settings} path="/settings" />
-              <Route component={AppContext} path="/apps/:id(\d+)" />
-              <AnonymousRoute component={EditPassword} exact path="/edit-password" />
-              <ProtectedRoute component={OrganizationInvite} exact path="/organization-invite" />
-              <Route component={VerifyEmail} exact path="/verify" />
-              <Route component={OAuth2Connect} exact path="/connect/:provider/callback" />
-              <AnonymousRoute component={Login} exact path="/login" />
-              {settings.enableRegistration && (
-                <AnonymousRoute component={Register} exact path="/register" />
-              )}
-              <Route component={ResetPassword} exact path="/reset-password" />
-              <Route component={EditPassword} exact path="/edit-password" />
-              <Route component={VerifyEmail} exact path="/verify" />
-              <Redirect to="/apps" />
-            </Switch>
-            <Message />
-          </ErrorHandler>
+          <OrganizationProvider>
+            <ErrorHandler fallback={ErrorFallback}>
+              <Helmet defaultTitle="Appsemble" titleTemplate="Appsemble · %s" />
+              <Toolbar />
+              <Switch>
+                <Route component={AppList} exact path="/apps" />
+                <ProtectedRoute component={Settings} path="/settings" />
+                <Route component={AppContext} path="/apps/:id(\d+)" />
+                <AnonymousRoute component={EditPassword} exact path="/edit-password" />
+                <ProtectedRoute component={OrganizationInvite} exact path="/organization-invite" />
+                <Route component={VerifyEmail} exact path="/verify" />
+                <Route component={OAuth2Connect} exact path="/connect/:provider/callback" />
+                <AnonymousRoute component={Login} exact path="/login" />
+                {settings.enableRegistration && (
+                  <AnonymousRoute component={Register} exact path="/register" />
+                )}
+                <Route component={ResetPassword} exact path="/reset-password" />
+                <Route component={EditPassword} exact path="/edit-password" />
+                <Route component={VerifyEmail} exact path="/verify" />
+                <Redirect to="/apps" />
+              </Switch>
+              <Message />
+            </ErrorHandler>
+          </OrganizationProvider>
         </UserProvider>
       </BrowserRouter>
     </IntlProvider>