+++ packages/studio/components/AppContext/AppContext.jsx
@@ -1,8 +1,10 @@
 import { Loader } from '@appsemble/react-components';
+import { permissions } from '@appsemble/utils';
 import PropTypes from 'prop-types';
 import React from 'react';
 import { Redirect, Route, Switch } from 'react-router-dom';
 
+import useOrganizations from '../../hooks/useOrganizations';
 import AppDetails from '../AppDetails';
 import AppSettings from '../AppSettings';
 import AppSideMenu from '../AppSideMenu';
@@ -15,39 +17,71 @@ import styles from './AppContext.css';
 /**
  * A wrapper which fetches the app definition and makes sure it is available to its children.
  */
-export default class AppContext extends React.Component {
-  static propTypes = {
-    getApp: PropTypes.func.isRequired,
-    match: PropTypes.shape().isRequired,
-    ready: PropTypes.bool.isRequired,
-  };
+function AppContext({ app = undefined, match, getApp, initAuth, ready }) {
+  React.useEffect(() => {
+    const initApp = async () => {
+      await initAuth();
+      await getApp(match.params.id);
+    };
 
-  async componentDidMount() {
-    const { match, getApp } = this.props;
-    await getApp(match.params.id);
+    initApp();
+  }, [getApp, initAuth, match.params.id]);
+
+  const organizations = useOrganizations();
+
+  if (!ready) {
+    return <Loader />;
   }
 
-  render() {
-    const { ready, match } = this.props;
+  const organization = organizations.find(org => org.id === app.OrganizationId);
 
-    if (!ready) {
-      return <Loader />;
-    }
-
-    return (
-      <div className={styles.container}>
-        <AppSideMenu />
-        <div className={styles.content}>
-          <Switch>
-            <Route component={AppDetails} exact path={match.path} />
-            <ProtectedRoute component={Editor} exact path={`${match.path}/edit`} />
-            <ProtectedRoute component={CMS} path={`${match.path}/resources`} />
-            <ProtectedRoute component={AppSettings} exact path={`${match.path}/settings`} />
-            <ProtectedRoute component={Notifications} exact path={`${match.path}/notifications`} />
-            <Redirect to={match.path} />
-          </Switch>
-        </div>
+  return (
+    <div className={styles.container}>
+      <AppSideMenu />
+      <div className={styles.content}>
+        <Switch>
+          <Route component={AppDetails} exact path={match.path} />
+          <ProtectedRoute
+            component={Editor}
+            exact
+            organization={organization}
+            path={`${match.path}/edit`}
+            permission={permissions.EditApps}
+          />
+          <ProtectedRoute
+            component={CMS}
+            organization={organization}
+            path={`${match.path}/resources`}
+            permission={permissions.EditApps}
+          />
+          <ProtectedRoute
+            component={AppSettings}
+            exact
+            organization={organization}
+            path={`${match.path}/settings`}
+            permission={permissions.EditAppSettings}
+          />
+          <ProtectedRoute
+            component={Notifications}
+            exact
+            organization={organization}
+            path={`${match.path}/notifications`}
+            permission={permissions.PushNotifications}
+          />
+          <Redirect to={match.path} />
+        </Switch>
       </div>
-    );
-  }
+    </div>
+  );
 }
+
+AppContext.propTypes = {
+  // eslint-disable-next-line react/require-default-props
+  app: PropTypes.shape(),
+  getApp: PropTypes.func.isRequired,
+  initAuth: PropTypes.func.isRequired,
+  match: PropTypes.shape().isRequired,
+  ready: PropTypes.bool.isRequired,
+};
+
+export default AppContext;