+++ packages/studio/hooks/useOrganizations.ts
@@ -1,0 +1,9 @@
+import { createContext, useContext } from 'react';
+
+import { Organization } from '../types';
+
+export const OrganizationContext = createContext<Organization[]>(null);
+
+export default function useOrganizations(): Organization[] {
+  return useContext(OrganizationContext);
+}