+++ packages/studio/types.ts
@@ -59,6 +59,21 @@ export interface OAuth2ClientCredentials {
   scopes: string[];
 }
 
+export type Role = 'Owner' | 'Maintainer' | 'AppEditor' | 'Member';
+
+export interface Member {
+  id: number;
+  name?: string;
+  primaryEmail?: string;
+  role: Role;
+}
+
+export interface Organization {
+  id: string;
+  name: string;
+  role: Role;
+}
+
 declare global {
   interface Window {
     settings: {